Add various display features and fix win issue
- Previously, there was an issue where the player
could still mark a grid cell even after the game
had stopped because a winner had been determined.
Fixed it by adding a check at the beginning and
end of the playTurn method

- Added a prompt to ask for player names and
displayed it on the page

- Added code to change the announcement elements
content, depending on whos turn it was and if
there was a winner.

- Removed a lot of console.logs previously used
for debugging